native function for conversion from BSON to a KjNode tree + performance traces if not outdeffed
