+++ src/lib/orionld/mongoCppLegacy/mongoCppLegacyEntityLookup.cpp
@@ -76,7 +76,7 @@ KjNode* mongoCppLegacyEntityLookup(const char* entityId)
     char*           title;
     char*           details;
 
-    kjTree = dbDataToKjTree(&bsonObj, &title, &details);
+    kjTree = dbDataToKjTree(&bsonObj, false, &title, &details);
     if (kjTree == NULL)
       LM_E(("%s: %s", title, details));
     break;
