+++ src/lib/orionld/mongoCppLegacy/mongoCppLegacyKjTreeFromBsonObj.cpp
@@ -40,109 +40,6 @@ extern "C"
 #include "orionld/common/orionldState.h"                             // orionldState, orionldStateDelayedFreeEnqueue
 #include "orionld/mongoCppLegacy/mongoCppLegacyKjTreeFromBsonObj.h"  // Own interface
 
-extern void         arrayToKjTree(KjNode* containerP, mongo::BSONArray* bsonArrayP, char** titleP, char** detailsP);
-extern const char*  mongoTypeName(mongo::BSONType type);
-
-
-
-void objectToKjTree(KjNode* containerP, mongo::BSONObj* bsonObjP, char** titleP, char** detailsP)
-{
-  for (mongo::BSONObj::iterator iter = bsonObjP->begin(); iter.more();)
-  {
-    mongo::BSONElement be       = iter.next();
-    mongo::BSONType    type     = be.type();
-    KjNode*            nodeP    = NULL;
-    const char*        nodeName = be.fieldName();
-
-    if (type == mongo::String)
-      nodeP = kjString(orionldState.kjsonP, nodeName, be.String().c_str());
-    else if (type == mongo::Bool)
-      nodeP = kjBoolean(orionldState.kjsonP, nodeName, be.Bool());
-    else if (type == mongo::NumberDouble)
-      nodeP = kjFloat(orionldState.kjsonP, nodeName, be.Number());
-    else if (type == mongo::jstNULL)
-      nodeP = kjNull(orionldState.kjsonP, nodeName);
-    else if (type == mongo::Object)
-    {
-      mongo::BSONObj bo = be.embeddedObject();
-      nodeP = kjObject(orionldState.kjsonP, nodeName);
-      objectToKjTree(nodeP, &bo, titleP, detailsP);
-    }
-    else if (type == mongo::Array)
-    {
-      mongo::BSONArray ba = (mongo::BSONArray) be.embeddedObject();
-      nodeP = kjArray(orionldState.kjsonP, nodeName);
-      arrayToKjTree(nodeP, &ba, titleP, detailsP);
-    }
-    else
-    {
-      LM_E(("Unsupported mongo type %d for field '%s'", type, nodeName));
-      continue;
-    }
-
-    kjChildAdd(containerP, nodeP);
-  }
-}
-
-
-void arrayToKjTree(KjNode* containerP, mongo::BSONArray* bsonArrayP, char** titleP, char** detailsP)
-{
-  for (mongo::BSONObj::iterator iter = bsonArrayP->begin(); iter.more();)
-  {
-    mongo::BSONElement be       = iter.next();
-    mongo::BSONType    type     = be.type();
-    KjNode*            nodeP    = NULL;
-    char*              nodeName = NULL;
-
-    if (type == mongo::String)
-      nodeP = kjString(orionldState.kjsonP, nodeName, be.String().c_str());
-    else if (type == mongo::Bool)
-      nodeP = kjBoolean(orionldState.kjsonP, nodeName, be.Bool());
-    else if (type == mongo::NumberDouble)
-      nodeP = kjFloat(orionldState.kjsonP, nodeName, be.Number());
-    else if (type == mongo::jstNULL)
-      nodeP = kjNull(orionldState.kjsonP, nodeName);
-    else if (type == mongo::Object)
-    {
-      mongo::BSONObj bo = be.embeddedObject();
-      nodeP = kjObject(orionldState.kjsonP, nodeName);
-      objectToKjTree(nodeP, &bo, titleP, detailsP);
-    }
-    else if (type == mongo::Array)
-    {
-      mongo::BSONArray ba = (mongo::BSONArray) be.embeddedObject();
-      nodeP = kjArray(orionldState.kjsonP, nodeName);
-      arrayToKjTree(nodeP, &ba, titleP, detailsP);
-    }
-    else
-    {
-      LM_E(("Unsupported mongo type %d for field '%s'", type, nodeName));
-      continue;
-    }
-
-    kjChildAdd(containerP, nodeP);
-  }
-}
-
-
-KjNode* mongoCppLegacyDataToKjTree(const void* dataP, bool isArray, char** titleP, char** detailsP)
-{
-  KjNode* rootP;
-
-  if (isArray == false)
-  {
-    rootP = kjObject(orionldState.kjsonP, NULL);
-    objectToKjTree(rootP, (mongo::BSONObj*) dataP, titleP, detailsP);
-  }
-  else
-  {
-    rootP = kjArray(orionldState.kjsonP, NULL);
-    arrayToKjTree(rootP, (mongo::BSONArray*) dataP, titleP, detailsP);
-  }
-
-  return rootP;
-}
-
 
 
 // -----------------------------------------------------------------------------
