+++ src/lib/orionld/mongoCppLegacy/mongoCppLegacySubscriptionGet.cpp
@@ -71,7 +71,7 @@ KjNode* mongoCppLegacySubscriptionGet(const char* subscriptionId)
     char*           title;
     char*           details;
 
-    kjTree = dbDataToKjTree(&bsonObj, &title, &details);
+    kjTree = dbDataToKjTree(&bsonObj, false, &title, &details);
     if (kjTree == NULL)
       LM_E(("%s: %s", title, details));
   }
